+++ config/initializers/prometheus.rb
@@ -4,7 +4,7 @@ require 'prometheus/client'
 Prometheus::Client.registry
 
 http_request_duration = Prometheus::Client::Histogram.new(
-  :http_request_duration_seconds,
+  :budget_app_http_request_duration_seconds,
   docstring: 'HTTP request duration in seconds',
   labels: %i[controller action method status]
 )
@@ -12,7 +12,7 @@ http_request_duration = Prometheus::Client::Histogram.new(
 begin
   Prometheus::Client.registry.register(http_request_duration)
 rescue Prometheus::Client::Registry::AlreadyRegisteredError
-  http_request_duration = Prometheus::Client.registry.get(:http_request_duration_seconds)
+  http_request_duration = Prometheus::Client.registry.get(:budget_app_http_request_duration_seconds)
 end
 
 ActiveSupport::Notifications.subscribe('process_action.action_controller') do |*args|